requirements.txt
streamlit
google-generativeai
python-dotenv
pyttsx3
pypiwin32
fastapi 
uvicorn
 pillow